% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_indicators.R
\name{addBBands}
\alias{addBBands}
\title{Add Bollinger Bands indicators
to the charts}
\usage{
addBBands(plot, cols = c("High", "Low", "Close"), ...)
}
\arguments{
\item{plot}{a kline, or OHLC, chart.}

\item{cols}{a vector of column names for the
Bollinger bands calculations.}

\item{...}{See \code{\link[TTR:bollingerBands]{TTR::BBands()}}}
}
\description{
Add Bollinger Bands indicators
to the charts
}
