% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_indicators.R
\name{addMACD}
\alias{addMACD}
\title{Add MACD indicator to the chart}
\usage{
addMACD(plot, ...)
}
\arguments{
\item{plot}{A plotly object of either
klines or OHLC}

\item{...}{See \code{\link[TTR:MACD]{TTR::MACD()}}}
}
\description{
Add MACD indicator to the chart
}
