% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_indicators.R
\name{addVlines}
\alias{addVlines}
\title{add vertical lines to
the chart}
\usage{
addVlines(plot, object, color = "steelblue")
}
\arguments{
\item{plot}{Chart}

\item{object}{an XTS object with dates
where the vertical line has to be drawn}

\item{color}{description}
}
\description{
A short description...

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
