% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_indicators.R
\name{addVolume}
\alias{addVolume}
\title{Add volume indicator
to the chart}
\usage{
addVolume(plot)
}
\arguments{
\item{plot}{A plotly object of either
klines or OHLC}
}
\description{
Add volume indicator
to the chart
}
