% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATOM.R
\docType{data}
\name{ATOM}
\alias{ATOM}
\title{USDT Denominated ATOM (ATOMUSDT) 15-Minute Intervals}
\format{
An \code{\link[xts:xts]{xts::xts()}} object with 97 rows and 5 columns structured as follows:
\describe{
\item{}{\strong{open} \link{numeric} Opening price.}
\item{}{\strong{high} \link{numeric} Highest price.}
\item{}{\strong{low} \link{numeric} Lowest price.}
\item{}{\strong{close} \link{numeric} Closing price.}
\item{}{\strong{volume} \link{numeric} Trading volume.}
}
}
\usage{
ATOM
}
\description{
This dataset contains time-series data for the ATOM cryptocurrency denominated in USDT (Tether),
captured in 15-minute intervals (OHLC-V format). The data spans from December 30 to December 31, 2023.
}
\examples{
# Load the dataset
data("ATOM")

# candlestick charts
chart(
  ticker = ATOM,
  main = kline(),
  sub = list(volume())
)
}
\keyword{datasets}
