% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableIntervals.R
\name{availableIntervals}
\alias{availableIntervals}
\title{Get available intervals}
\usage{
availableIntervals(source = "binance", type = "ohlc", futures = TRUE)
}
\arguments{
\item{source}{A \link{character}-vector of \link{length} 1. \code{binance} by default. See \code{\link[=available_exchanges]{available_exchanges()}} for available exchanges.}

\item{type}{\link{character}-vector of length 1. See details.}

\item{futures}{A \link{logical}-vector of \link{length} 1. \link{TRUE} by default. Returns futures market if \link{TRUE}, spot market otherwise.}
}
\value{
An \code{\link[=invisible]{invisible()}} \link{character} vector containing the  available intervals on
the exchange, market and endpoint
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\subsection{Available types}{
\itemize{
\item \code{"ohlc"} <\link{character}> Open, High, Low, Close and Volume market data for the \code{\link[=get_quote]{get_quote()}}-function
\item \code{"lsratio"} <\link{character}> Long-Short ratios for the \code{\link[=get_lsratio]{get_lsratio()}}-function
\item \code{"fundingrate"} <\link{character}> Funding rates for the \code{\link[=get_fundingrate]{get_fundingrate()}}-function
\item \code{"interest"} <\link{character} >Open interest on perpetual contracts on both sides for the \code{\link[=get_openinterest]{get_openinterest()}}-function
}
}

\subsection{Limits}{

The endpoints supported by the \code{\link[=available_exchanges]{available_exchanges()}} are not uniform,
so exchanges available for, say, \code{\link[=get_lsratio]{get_lsratio()}} is not necessarily the same as those available for \code{\link[=get_quote]{get_quote()}}
}
}
\examples{
\dontrun{
  # script start;

  # available intervals
  # at kucoin futures market
  cryptoQuotes::available_intervals(
    source  = 'kucoin',
    futures = TRUE,
    type    = "ohlc"
  )

  # available intervals
  # at kraken spot market
  cryptoQuotes::available_intervals(
    source  = 'kraken',
    futures = FALSE,
    type    = "ohlc"
  )

  # script end;
}
}
\seealso{
Other deprecated: 
\code{\link{availableExchanges}()},
\code{\link{availableTickers}()},
\code{\link{getFGIndex}()},
\code{\link{getLSRatio}()},
\code{\link{getQuote}()}
}
\author{
Serkan Korkmaz
}
\concept{deprecated}
