% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTickers.R
\name{availableTickers}
\alias{availableTickers}
\title{availableTickers}
\usage{
availableTickers(source = "binance", futures = TRUE)
}
\arguments{
\item{source}{a character vector of length 1. The source of the API}

\item{futures}{a logical value. Default TRUE.}
}
\value{
Returns a character vector
of length N equal to the tradable tickers
}
\description{
This function returns all available
pairs on the exchanges.
}
