% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTCUSDT.R
\docType{data}
\name{BTCUSDT}
\alias{BTCUSDT}
\title{USDT denominated Bitcoin(BTC) with 1 week intervals}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 99 rows and 5 columns.
}
\usage{
BTCUSDT
}
\description{
A xts object with weekly OHLCV of USDT denominated Bitcoin with 99 rows and 5 columns, from 2022-02-07 to 2023-12-25.
}
\details{
\describe{
\item{Open}{Opening price}
\item{High}{Highest price}
\item{Low}{Lowest price}
\item{Close}{Closing price}
\item{Volume}{Volume}
}
}
\seealso{
Other data: 
\code{\link{ATOMUSDT}},
\code{\link{DOGEUSDT}},
\code{\link{FGIndex}}
}
\concept{data}
\keyword{datasets}
