% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Create a candlestick chart
with volume and bollinger bands}
\usage{
chart(quote, deficiency = FALSE, slider = TRUE)
}
\description{
This function returns a plotly candlestick with
the most common indicators.
}
