% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{chart_layout}
\alias{chart_layout}
\title{Create a list of layout elements on subcharts}
\usage{
chart_layout(x, layout_element, layout_attribute)
}
\arguments{
\item{x}{\link{integer}-vector of \link{length} 1.}

\item{layout_element}{\link{character}-vector of \link{length} 1.
\link[plotly:layout]{plotly::layout} elements. See example.}

\item{layout_attribute}{\link{character}-vector of \link{length} 1.
\link[plotly:layout]{plotly::layout} element value. See example.}
}
\value{
A \link{list} of layout elements.
}
\description{
Create a list of layout elements on subcharts
}
\examples{
\dontrun{
chart_layout(
  x = 1:plot_list_length,
  layout_element = "yaxis",
  layout_attribute = list(
  gridcolor = if (dark) "#40454c" else  '#D3D3D3' # Was CCCCCC
    )
)
}


}
\seealso{
Other development tools: 
\code{\link{GET}()},
\code{\link{convert_date}()},
\code{\link{default_dates}()},
\code{\link{fetch}()},
\code{\link{flatten}()},
\code{\link{is.date}()}
}
\concept{development tools}
\keyword{internal}
