% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convert_date}
\alias{convert_date}
\title{Convert dates passed to UNIX}
\usage{
convert_date(x, multiplier)
}
\arguments{
\item{x}{a \link{numeric} vector, or \link{date}-type object}

\item{multiplier}{\link{numeric}}
}
\value{
A vector of same length as x.
}
\description{
This function converts dates to UNIX time if passed to the API, and converts
it to \link{POSIXct} from API
}
\details{
If x is numeric, then the function assumes
that its a return value
}
\seealso{
Other development tools: 
\code{\link{GET}()},
\code{\link{chart_layout}()},
\code{\link{default_dates}()},
\code{\link{fetch}()},
\code{\link{flatten}()},
\code{\link{is.date}()}
}
\concept{development tools}
\keyword{internal}
