% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptoQuotes-package.R
\docType{package}
\name{cryptoQuotes-package}
\alias{cryptoQuotes}
\alias{cryptoQuotes-package}
\title{cryptoQuotes: Access Market Data From Major Cryptocurrency Exchanges}
\description{
Access spot and futures markets from major cryptocurrency exchanges.
}
\author{
\strong{Maintainer}: Serkan Korkmaz \email{serkor1@duck.com} [contributor]

}
\keyword{internal}
