% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{default_dates}
\alias{default_dates}
\title{Get the minimum and maximum date range}
\usage{
default_dates(interval, from = NULL, to = NULL, length = 200, limit = NULL)
}
\arguments{
\item{interval}{A \link{character}-vector of \link{length} 1. \verb{1d} by default.
See \code{\link[=available_intervals]{available_intervals()}} for available intervals.}

\item{from}{An optional \link{character}-, \link{date}- or
\link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}

\item{to}{An optional \link{character}-, \link{date}- or
\link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}

\item{length}{a \link{numeric}-value of \link{length} 1.
The desired distance between \code{from} and \code{to}.}
}
\value{
A vector of minimum and maximum dates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

All \code{\link[=available_exchanges]{available_exchanges()}} have different limitations on the
returned data this function is adaptive in nature so enforces
compliance on the limits
}
\seealso{
Other development tools: 
\code{\link{GET}()},
\code{\link{chart_layout}()},
\code{\link{convert_date}()},
\code{\link{fetch}()},
\code{\link{flatten}()},
\code{\link{is.date}()}
}
\author{
Serkan Korkmaz
}
\concept{development tools}
\keyword{internal}
