% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOGE.R
\docType{data}
\name{DOGE}
\alias{DOGE}
\title{USDT denominated DOGECOIN in 1m intervals}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 61 rows and 5 columns.
}
\usage{
DOGE
}
\description{
A xts object with 1m OHLCV of USDT denominated Dogecoin with 61 rows and 5 columns.
}
\details{
\describe{
\item{Open}{Opening price}
\item{High}{Highest price}
\item{Low}{Lowest price}
\item{Close}{Closing price}
\item{Volume}{Volume}
}
}
\seealso{
Other data: 
\code{\link{ATOM}},
\code{\link{BTC}},
\code{\link{FGIndex}}
}
\concept{data}
\keyword{datasets}
