% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{flatten}
\alias{flatten}
\title{flatten nested lists}
\usage{
flatten(x)
}
\arguments{
\item{x}{A \link{list}}
}
\value{
An unnested \link{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Flatten a nested \link{list}, regardless of its level of nesting.
}
\examples{
# script start;

# 1) create a nested list
nested_list <- list(
  a = 1,
  b = list(
    c = 2,
    d = 3
  )
)

# 2) flatten the
# nested list
cryptoQuotes:::flatten(
  nested_list
)

# script end;
}
\seealso{
Other development tools: 
\code{\link{GET}()},
\code{\link{chart_layout}()},
\code{\link{convert_date}()},
\code{\link{default_dates}()},
\code{\link{fetch}()},
\code{\link{is.date}()}
}
\concept{development tools}
\keyword{internal}
