% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuote.R
\name{getQuote}
\alias{getQuote}
\title{getQuote}
\usage{
getQuote(
  ticker,
  source = "binance",
  futures = TRUE,
  interval,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{ticker}{A character vector of length 1. Has to be denominated in some currency}

\item{source}{A character vector of length 1, of the specific}

\item{interval}{A character vector of length 1, of desired intervals.}
}
\description{
This function returns
a crypto quote
}
