% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is.date}
\alias{is.date}
\title{Check if values are valid dates}
\usage{
is.date(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\link{TRUE} if its either POSIXct, POSIXt or Date. \link{FALSE} otherwise.
}
\description{
This function check is equivalent to \code{\link[=is.numeric]{is.numeric()}}, \code{\link[=is.logical]{is.logical()}},
and checks for the date type classes POSIXct, POSIXt and Date.
And wether the character vector can be formatted to dates.
}
\seealso{
Other development tools: 
\code{\link{GET}()},
\code{\link{chart_layout}()},
\code{\link{convert_date}()},
\code{\link{default_dates}()},
\code{\link{fetch}()},
\code{\link{flatten}()}
}
\concept{development tools}
\keyword{internal}
