% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_quote.R
\name{pline}
\alias{pline}
\title{Line Chart}
\usage{
pline(price = "close", ...)
}
\arguments{
\item{price}{A \link{character}-vector of \link{length} 1. "close" by default.}

\item{...}{For internal use. Please ignore.}
}
\value{
An \link{invisible} \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A high-level \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-function for charting
Open, High, Low and Close prices.
}
\examples{
# script start;

# Charting BTC using
# line charts with closing price
# as main chart
cryptoQuotes::chart(
  ticker = BTC,
  main   = cryptoQuotes::pline(),
  sub    = list(
    cryptoQuotes::volume()
  )
)

# script end;
}
\seealso{
Other price charts: 
\code{\link{chart}()},
\code{\link{kline}()},
\code{\link{ohlc}()}
}
\author{
Serkan Korkmaz
}
\concept{price charts}
