% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN.R
\name{removeBound}
\alias{removeBound}
\title{remove upper and lower bounds
from an XTS object}
\usage{
removeBound(xts, bounds = c("upper"))
}
\arguments{
\item{xts}{A xts-object that needs its bounds modified.}

\item{bounds}{A character vector of length 1. Has to be one of \code{c('upper','lower','both')}. Defaults to Upper.}
}
\value{
Returns an xts-class object with its bounds removed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link[stats:window]{stats::window()}}-function has inclusive upper and lower bounds,
which in some cases is an undesirable feature. This high level function
removes the bounds if desired
}
\seealso{
Other convinience: 
\code{\link{calibrateWindow}()},
\code{\link{splitWindow}()}
}
\concept{convinience}
