% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN.R
\name{splitWindow}
\alias{splitWindow}
\title{split xts object iteratively in lists of desired intervals}
\usage{
splitWindow(xts, by, bounds = "upper")
}
\arguments{
\item{xts}{A xts-object that needs needs to be split.}

\item{by}{A reference \code{\link[zoo:index]{zoo::index()}}-object, to be split by.}

\item{bounds}{A character vector of length 1. Has to be one of \code{c('upper','lower','both')}. Defaults to Upper.}
}
\value{
Returns a list of iteratively restricted xts objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link[=splitWindow]{splitWindow()}}-function is a high level wrapper of the \code{\link[stats:window]{stats::window()}}-function which restricts
the intervals between the first and second index value iteratively
}
\seealso{
Other convinience: 
\code{\link{calibrateWindow}()},
\code{\link{removeBound}()}
}
\concept{convinience}
