% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_indicators.R
\name{addMA}
\alias{addMA}
\title{Add various Moving Average indicators
to the chart.}
\usage{
addMA(plot, FUN = TTR::SMA, ...)
}
\arguments{
\item{plot}{A plotly object of either
klines or OHLC}

\item{FUN}{A named function calculating MAs. See \code{\link[TTR:MACD]{TTR::MACD()}}}

\item{...}{See \code{\link[TTR:MACD]{TTR::MACD()}}}
}
\description{
Add various Moving Average indicators
to the chart.
}
