% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_indicators.R
\name{addRSI}
\alias{addRSI}
\title{Add RSI indicators to your
chart}
\usage{
addRSI(plot, ...)
}
\arguments{
\item{plot}{A plotly object of either
klines or OHLC}

\item{...}{See \code{\link[TTR:RSI]{TTR::RSI()}}}
}
\description{
A short description...

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
