% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_indicators.R
\name{addVlines}
\alias{addVlines}
\title{add vertical lines to
the chart}
\usage{
addVlines(plot, object, color = "steelblue")
}
\arguments{
\item{plot}{Chart}

\item{object}{an XTS object with dates
where the vertical line has to be drawn}

\item{color}{description}
}
\description{
A short description...

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# script: scr_addVlines
# date: 2023-10-25
# author: Serkan Korkmaz, serkor1@duck.com
# objective: Describe the usage
# of addVlines
# script start;

# laod library
library(cryptoQuotes)

# 1) add random a random
# buy date to plot
ATOMUSDT$buy <- sample(
  c(1, rep(0, nrow(ATOMUSDT) - 1))
)

# 2) subset and
# chart
chart(
  chart = kline(
    ATOMUSDT
  ) \%>\% addVlines(
    subset(
      ATOMUSDT,
      buy == 1
    )
  )
)

# script end;
}
