% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATOM.R
\docType{data}
\name{ATOM}
\alias{ATOM}
\title{USDT denominated ATOMS with 15m intervals}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 97 rows and 5 columns.
}
\usage{
ATOM
}
\description{
A xts object with 15m OHLCV of USDT denominated ATOM with 97 rows and 5 columns, from 2023-12-30 to 2023-12-31.
}
\details{
\describe{
\item{Open}{Opening price}
\item{High}{Highest price}
\item{Low}{Lowest price}
\item{Close}{Closing price}
\item{Volume}{Volume}
}
}
\seealso{
Other data: 
\code{\link{BTC}},
\code{\link{DOGE}},
\code{\link{FGIndex}}
}
\concept{data}
\keyword{datasets}
