% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATOM.R
\docType{data}
\name{ATOM}
\alias{ATOM}
\title{USDT Denominated ATOM (ATOMUSDT) 15-Minute Intervals}
\format{
An \code{\link[xts:xts]{xts::xts()}}-object with 97 rows and 5 columns,

\describe{
\item{index}{<\link{POSIXct}> The time-index}
\item{open}{<\link{numeric}> Opening price}
\item{high}{<\link{numeric}> Highest price}
\item{low}{<\link{numeric}> Lowest price}
\item{close}{<\link{numeric}> Closing price}
\item{volume}{<\link{numeric}> Trading volume}
}
}
\usage{
ATOM
}
\description{
This dataset contains time-series data for the ATOM (ATOM)
denominated in USDT (Tether), captured in 15-minute intervals.
The data spans from December 30 to December 31, 2023.
}
\examples{
# Load the dataset
data("ATOM")

# chart
chart(
  ticker = ATOM,
  main = kline(),
  sub = list(volume())
)
}
\keyword{datasets}
