% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATOMUSDT.R
\docType{data}
\name{ATOMUSDT}
\alias{ATOMUSDT}
\title{USDT denominated ATOMS, with 15m intervals}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 97 rows and 5 columns.
}
\usage{
ATOMUSDT
}
\description{
A xts object with 15m OHLC of USDT denominated ATOM with 97 rows and 4 columns, from 2023-01-01 to 2023-01-02:
\describe{
\item{Open}{Opening price}
\item{High}{Highest price}
\item{Low}{Lowest price}
\item{Close}{Closing price}
}
}
\keyword{datasets}
