% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableExchanges.R
\name{availableExchanges}
\alias{availableExchanges}
\title{Get available exchanges}
\usage{
availableExchanges()
}
\value{
Invisbly returns a character vector.
}
\description{
This function returns all
available exchanges as a message in the console.
}
\examples{
# script:
# date: 2023-10-06
# author: Serkan Korkmaz, serkor1@duck.com
# objective:
# script start;

## return all
## available exchanges
cryptoQuotes::availableExchanges()

# script end;
}
