% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableIntervals.R
\name{availableIntervals}
\alias{availableIntervals}
\title{availableIntervals}
\usage{
availableIntervals(source = "binance", futures = TRUE)
}
\arguments{
\item{source}{character vector of length one. Must be the name of the
supported exchange}

\item{futures}{logical. TRUE by default. If FALSE, spotmarket are
returned}
}
\description{
This function shows all
available intervals available
from each exchange
}
