% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableIntervals.R
\name{availableIntervals}
\alias{availableIntervals}
\title{See all available intervals for the futures and spot markets
on the desired exchange}
\usage{
availableIntervals(source = "binance", futures = TRUE)
}
\arguments{
\item{source}{character vector of length one. Must be the name of the
supported exchange}

\item{futures}{logical. TRUE by default. If FALSE, spot market are
returned}
}
\value{
Invisbly returns a character vector.
}
\description{
This function shows all
available intervals available
from each exchange
}
\examples{
# script:
# date: 2023-10-06
# author: Serkan Korkmaz, serkor1@duck.com
# objective:
# script start;

# available intervals
# at kucoin futures market
cryptoQuotes::availableIntervals(
  source = 'kucoin',
  futures = TRUE
)

# available intervals
# at kraken spot market
cryptoQuotes::availableIntervals(
  source = 'kraken',
  futures = FALSE
)

# script end;
}
