% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTC.R
\docType{data}
\name{BTC}
\alias{BTC}
\title{USDT denominated Bitcoin(BTC) with 1 week intervals}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 52 rows and 5 columns.
}
\usage{
BTC
}
\description{
A xts object with weekly OHLCV of USDT denominated Bitcoin with 52 rows and 5 columns, from 2023-01-01 to 2023-12-31.
}
\details{
\describe{
\item{Open}{Opening price}
\item{High}{Highest price}
\item{Low}{Lowest price}
\item{Close}{Closing price}
\item{Volume}{Volume}
}
}
\seealso{
Other data: 
\code{\link{ATOM}},
\code{\link{DOGE}},
\code{\link{FGIndex}}
}
\concept{data}
\keyword{datasets}
