% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTC.R
\docType{data}
\name{BTC}
\alias{BTC}
\title{USDT Denominated Bitcoin (BTCUSDT) Weekly Intervals}
\format{
An \code{\link[xts:xts]{xts::xts()}} object structured as follows:
\describe{
\item{open}{\link{numeric} Opening price of Bitcoin in USDT at the start of each week.}
\item{high}{\link{numeric} Highest price of Bitcoin in USDT during the week.}
\item{low}{\link{numeric} Lowest price of Bitcoin in USDT during the week.}
\item{close}{\link{numeric} Closing price of Bitcoin in USDT at the end of each week.}
\item{volume}{\link{numeric} Trading volume of Bitcoin for the week, measured in units of Bitcoin.}
}
}
\usage{
BTC
}
\description{
This dataset contains time-series data for Bitcoin (BTC) denominated in USDT (Tether),
captured in weekly intervals (OHLC-V format). The dataset consists of 52 rows and 5 columns,
covering the period from January 1, 2023, to December 31, 2023.
}
\examples{
# Load the dataset
data("BTC")

# candlestick charts
chart(
  ticker = BTC,
  main = kline(),
  sub = list(volume())
)
}
\keyword{datasets}
