% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN.R
\name{calibrateWindow}
\alias{calibrateWindow}
\title{calibrate the time window of a list of xts objects}
\usage{
calibrateWindow(list, FUN, ...)
}
\arguments{
\item{list}{A list of xts objects.}

\item{FUN}{A function applied to each element of the
list}

\item{...}{optional arguments passed to \code{FUN}.}
}
\value{
Returns a xts object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is a high-level wrapper of \link{do.call} and \link{lapply} which
modifies each xts object stored in a \code{\link[=list]{list()}}.
}
\seealso{
Other convinience: 
\code{\link{removeBound}()},
\code{\link{splitWindow}()}
}
\concept{convinience}
