% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Chart your data}
\usage{
chart(chart, slider = TRUE)
}
\arguments{
\item{chart}{A kline or ohlc chart with and without
indicators}

\item{slider}{Logical. TRUE by default.}
}
\description{
Chart your data
}
