% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Create a candlestick chart
with volume, bollinger bands and moving average convergence and divergence
indicators}
\usage{
chart(quote, deficiency = FALSE)
}
\description{
This function returns a plotly candlestick with
the most common indicators.
}
