% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptoQuotes-package.R
\docType{package}
\name{cryptoQuotes-package}
\alias{cryptoQuotes}
\alias{cryptoQuotes-package}
\title{cryptoQuotes: A streamlined access to OHLC-V market data and sentiment indicators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This high-level API client offers a streamlined access to comprehensive cryptocurrency market data from major exchanges. It features robust OHLC-V (Open, High, Low, Close, Volume) candle data with flexible granularity, ranging from seconds to months, and includes insightful sentiment indicators. By aggregating data directly from leading exchanges, this package ensures a reliable and stable flow of market information, eliminating the need for complex, low-level API interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://serkor1.github.io/cryptoQuotes/}
  \item \url{https://github.com/serkor1/cryptoQuotes}
  \item Report bugs at \url{https://github.com/serkor1/cryptoQuotes/issues}
}

}
\author{
\strong{Maintainer}: Serkan Korkmaz \email{serkor1@duck.com} (\href{https://orcid.org/0000-0002-5052-0982}{ORCID}) [contributor, copyright holder]

Other contributors:
\itemize{
  \item Jonas Cuzulan Hirani \email{jjh@vive.dk} (\href{https://orcid.org/0000-0002-9512-1993}{ORCID}) [contributor]
}

}
\keyword{internal}
