% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptoQuotes-package.R
\docType{package}
\name{cryptoQuotes-package}
\alias{cryptoQuotes}
\alias{cryptoQuotes-package}
\title{cryptoQuotes: Access OHLC Market Data from Major Cryptocurrency Exchanges}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A high level library to extract cryptocurrency OHLC market data from major centralized exchanges. The library supports all available intervals supported by the exchange API for spot and perpetual futures markets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://serkor1.github.io/cryptoQuotes/}
  \item \url{https://github.com/serkor1/cryptoQuotes}
  \item Report bugs at \url{https://github.com/serkor1/cryptoQuotes/issues}
}

}
\author{
\strong{Maintainer}: Serkan Korkmaz \email{serkor1@duck.com} (\href{https://orcid.org/0000-0002-5052-0982}{ORCID}) [contributor, copyright holder]

}
\keyword{internal}
