% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptoQuotes-package.R
\docType{package}
\name{cryptoQuotes-package}
\alias{cryptoQuotes}
\alias{cryptoQuotes-package}
\title{cryptoQuotes: Open Access to Cryptocurrency Market Data, Sentiment Indicators and Interactive Charts}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This high-level API client provides open access to cryptocurrency market data, sentiment indicators, and interactive charting tools. The data is sourced from major cryptocurrency exchanges via 'curl' and returned in 'xts'-format. The data comes in open, high, low, and close (OHLC) format with flexible granularity, ranging from seconds to months. This flexibility makes it ideal for developing and backtesting trading strategies or conducting detailed market analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://serkor1.github.io/cryptoQuotes/}
  \item \url{https://github.com/serkor1/cryptoQuotes}
  \item Report bugs at \url{https://github.com/serkor1/cryptoQuotes/issues}
}

}
\author{
\strong{Maintainer}: Serkan Korkmaz \email{serkor1@duck.com} (\href{https://orcid.org/0000-0002-5052-0982}{ORCID}) [contributor, copyright holder]

Other contributors:
\itemize{
  \item Jonas Cuzulan Hirani \email{jjh@vive.dk} (\href{https://orcid.org/0000-0002-9512-1993}{ORCID}) [contributor]
}

}
\keyword{internal}
