% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptoQuotes-package.R
\docType{package}
\name{cryptoQuotes-package}
\alias{cryptoQuotes}
\alias{cryptoQuotes-package}
\title{cryptoQuotes: A Streamlined Access to Cryptocurrency OHLC-V Market Data and Sentiment Indicators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This high-level API client offers a streamlined access to public cryptocurrency market data and sentiment indicators. It features OHLC-V (Open, High, Low, Close, Volume) that comes with granularity ranging from seconds to months and essential sentiment indicators to develop and backtest trading strategies, or conduct detailed market analysis. By interacting directly with the major cryptocurrency exchanges this package ensures a reliable, and stable, flow of market information, eliminating the need for complex, low-level API interactions or webcrawlers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://serkor1.github.io/cryptoQuotes/}
  \item \url{https://github.com/serkor1/cryptoQuotes}
  \item Report bugs at \url{https://github.com/serkor1/cryptoQuotes/issues}
}

}
\author{
\strong{Maintainer}: Serkan Korkmaz \email{serkor1@duck.com} (\href{https://orcid.org/0000-0002-5052-0982}{ORCID}) [contributor, copyright holder]

Other contributors:
\itemize{
  \item Jonas Cuzulan Hirani \email{jjh@vive.dk} (\href{https://orcid.org/0000-0002-9512-1993}{ORCID}) [contributor]
}

}
\keyword{internal}
