% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOGE.R
\docType{data}
\name{DOGE}
\alias{DOGE}
\title{USDT Denominated DOGECOIN (DOGECOINUSDT) 1-Minute Intervals}
\format{
An \code{\link[xts:xts]{xts::xts()}} object structured as follows:
\describe{
\item{}{\strong{open} \link{numeric} Opening price.}
\item{}{\strong{high} \link{numeric} Highest price.}
\item{}{\strong{low} \link{numeric} Lowest price.}
\item{}{\strong{close} \link{numeric} Closing price.}
\item{}{\strong{volume} \link{numeric} Trading volume.}
}
}
\usage{
DOGE
}
\description{
This dataset contains time-series data for the DOGECOIN cryptocurrency denominated in USDT (Tether),
captured in 1-minute intervals (OHLC-V format). The dataset consists of 61 rows and 5 columns and spans 2022-01-14 07:00:00 CET to 2022-01-14 08:00:00 CET.
}
\examples{
# Load the dataset
data("DOGE")

# Plot candlestick charts to visualize price movements
chart(
  ticker = DOGE,
  main = kline(),
  sub = list(volume())
)
}
\keyword{datasets}
