% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGIndex.R
\docType{data}
\name{FGIndex}
\alias{FGIndex}
\title{Fear and Greed Index Values}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 364 rows and 1 columns.
}
\usage{
FGIndex
}
\description{
A xts object with Fear and Greed Index value. It has 689 rows, and 1 colum. Extracted from 2023-01-01 to 2023-12-31
}
\details{
\describe{
\item{FGI}{Daily Fear and Greed Index Value}
}
}
\seealso{
Other data: 
\code{\link{ATOM}},
\code{\link{BTC}},
\code{\link{DOGE}}
}
\concept{data}
\keyword{datasets}
