% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGIndex.R
\docType{data}
\name{FGIndex}
\alias{FGIndex}
\title{Fear and Greed Index Values (FGIndex) Daily Intervals}
\format{
An \code{\link[xts:xts]{xts::xts()}} object structured as follows:
\describe{
\item{}{\strong{fgi}(\link{numeric}) The daily fear and greed index value}
}
}
\usage{
FGIndex
}
\description{
This dataset contains daily values of the Fear and Greed Index for the year 2023, which is used to measure the sentiments of investors in the market. The data spans from January 1, 2023, to December 31, 2023.
}
\details{
The Fear and Greed Index goes from 0-100, and can be classified as follows,

\describe{
\item{}{\strong{0-24} Extreme Fear.}
\item{}{\strong{25-44} Fear.}
\item{}{\strong{45-55} Neutral.}
\item{}{\strong{56-75}  Greed.}
\item{}{\strong{76-100} Extreme Greed.}
}
}
\examples{
# Load the dataset
data("FGIndex")

# Get a summary of index values
summary(FGIndex)
}
\keyword{datasets}
