% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fgi.R
\name{get_fgindex}
\alias{get_fgindex}
\title{Get the daily Fear and Greed Index
in the cryptocurrency market}
\usage{
get_fgindex(from = NULL, to = NULL)
}
\arguments{
\item{from}{An optional \link{character}, \link{date} or \link{POSIXct} vector of \link{length} 1. \link{NULL} by default.}

\item{to}{An optional \link{character}, \link{date} or \link{POSIXct} vector of \link{length} 1. \link{NULL} by default.}
}
\value{
An \link{xts}-object containing,

\describe{
\item{}{<\link{POSIXct}> \strong{\code{index}:} The time-index}
\item{}{<\link{numeric}> \strong{\code{fgi}:} The daily fear and greed index value}
}

\strong{Output}

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>                     fgi
#> 2024-05-02 02:00:00  43
#> 2024-05-03 02:00:00  48
#> 2024-05-04 02:00:00  67
#> 2024-05-05 02:00:00  69
#> 2024-05-06 02:00:00  71
#> 2024-05-07 02:00:00  68
}\if{html}{\out{</div>}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get the daily fear and greed index.
}
\details{
\subsection{Classification}{

The Fear and Greed Index goes from 0-100, and can be classified as follows,

\describe{
\item{}{\strong{0-24:} Extreme Fear.}
\item{}{\strong{25-44:} Fear.}
\item{}{\strong{45-55:} Neutral.}
\item{}{\strong{56-75:} Greed.}
\item{}{\strong{76-100:}  Extreme Greed.}
}
}

\subsection{About the Fear and Greed Index}{

The fear and greed index is a market sentiment indicator that measures investor emotions to
gauge whether they are generally fearful (indicating potential selling pressure) or greedy (indicating potential buying enthusiasm)
}
}
\examples{
\dontrun{
  # script start;

  # 1) get the fear and greed index
  # for the last 7 days
  tail(
    fgi <- cryptoQuotes::get_fgindex(
      from = Sys.Date() - 7
    )
  )

  # script end;
}

}
\seealso{
Other get-function: 
\code{\link{get_fundingrate}()},
\code{\link{get_lsratio}()},
\code{\link{get_openinterest}()},
\code{\link{get_quote}()}
}
\author{
Serkan Korkmaz
}
\concept{get-function}
