% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLSRatio.R
\name{getLSRatio}
\alias{getLSRatio}
\title{Get the long to short ratio of a cryptocurrency pair}
\usage{
getLSRatio(
  ticker,
  interval = "1d",
  source = "binance",
  from = NULL,
  to = NULL,
  top = FALSE
)
}
\arguments{
\item{ticker}{A \link{character}-vector of \link{length} 1. See \code{\link[=available_tickers]{available_tickers()}} for available tickers.}

\item{interval}{A \link{character}-vector of \link{length} 1. \verb{1d} by default. See \code{\link[=available_intervals]{available_intervals()}} for available intervals.}

\item{source}{A \link{character}-vector of \link{length} 1. \code{binance} by default. See \code{\link[=available_exchanges]{available_exchanges()}} for available exchanges.}

\item{from}{An optional \link{character}-, \link{date}- or \link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}

\item{to}{An optional \link{character}-, \link{date}- or \link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}

\item{top}{A \link{logical} vector. \link{FALSE} by default. If \link{TRUE} it returns the top traders Long-Short ratios.}
}
\value{
An \link{xts}-object containing,

\item{index}{<\link{POSIXct}> the time-index}
\item{long}{<\link{numeric}> the share of longs}
\item{short}{<\link{numeric}> the share of shorts}
\item{ls_ratio}{<\link{numeric}> the ratio of longs to shorts}

\strong{Sample output}

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>                       long  short ls_ratio
#> 2024-05-12 02:00:00 0.5656 0.4344 1.302026
#> 2024-05-13 02:00:00 0.5635 0.4365 1.290951
#> 2024-05-14 02:00:00 0.5447 0.4553 1.196354
#> 2024-05-15 02:00:00 0.5616 0.4384 1.281022
#> 2024-05-16 02:00:00 0.5390 0.4610 1.169197
#> 2024-05-17 02:00:00 0.5416 0.4584 1.181501
}\if{html}{\out{</div>}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Get the long-short ratio for any \code{\link[=available_tickers]{available_tickers()}} from the \code{\link[=available_exchanges]{available_exchanges()}}
}
\details{
\subsection{On time-zones and dates}{

Values passed to \code{from} or \code{to} must be coercible by \code{\link[=as.Date]{as.Date()}}, or \code{\link[=as.POSIXct]{as.POSIXct()}}, with a format of either \code{"\%Y-\%m-\%d"} or \code{"\%Y-\%m-\%d \%H:\%M:\%S"}. By default
all dates are passed and returned with \code{\link[=Sys.timezone]{Sys.timezone()}}.
}

\subsection{On returns}{

If only \code{from} is provided 200 pips are returned up to \code{Sys.time()}. If only \code{to} is provided 200 pips up to the specified date
is returned.
}
}
\examples{
\dontrun{
  # script start;

  LS_BTC <- cryptoQuotes::get_lsratio(
    ticker   = 'BTCUSDT',
    interval = '15m',
    from     = Sys.Date() - 1,
    to       = Sys.Date()
  )

  # end of scrtipt;
}

}
\seealso{
Other deprecated: 
\code{\link{availableExchanges}()},
\code{\link{availableIntervals}()},
\code{\link{availableTickers}()},
\code{\link{getFGIndex}()},
\code{\link{getQuote}()}
}
\author{
Jonas Cuzulan Hirani
}
\concept{deprecated}
