% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLSRatio.R
\name{getLSRatio}
\alias{getLSRatio}
\title{Get the long to short ratio of a cryptocurrency pair}
\usage{
getLSRatio(
  ticker,
  interval = "1d",
  source = "binance",
  from = NULL,
  to = NULL,
  top = FALSE
)
}
\arguments{
\item{ticker}{A \link{character}-vector of \link{length} 1. See \code{\link[=available_tickers]{available_tickers()}} for available tickers.}

\item{interval}{A \link{character}-vector of \link{length} 1. \verb{1d} by default. See \code{\link[=available_intervals]{available_intervals()}} for available intervals.}

\item{source}{A \link{character}-vector of \link{length} 1. \code{binance} by default. See \code{\link[=available_exchanges]{available_exchanges()}} for available exchanges.}

\item{from}{An optional \link{character}-, \link{date}- or \link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}

\item{to}{An optional \link{character}-, \link{date}- or \link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}

\item{top}{A \link{logical} vector. \link{FALSE} by default. If \link{TRUE} it returns the top traders Long-Short ratios.}
}
\value{
An \link{xts}-object containing,

\item{index}{<\link{POSIXct}> the time-index}
\item{long}{<\link{numeric}> the share of longs}
\item{short}{<\link{numeric}> the share of shorts}
\item{ls_ratio}{<\link{numeric}> the ratio of longs to shorts}

\strong{Sample output}

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>                       long  short ls_ratio
#> 2024-05-07 02:00:00 0.5492 0.4508 1.218279
#> 2024-05-08 02:00:00 0.5467 0.4533 1.206045
#> 2024-05-09 02:00:00 0.5580 0.4420 1.262443
#> 2024-05-10 02:00:00 0.5631 0.4369 1.288853
#> 2024-05-11 02:00:00 0.5766 0.4234 1.361833
#> 2024-05-12 02:00:00 0.5656 0.4344 1.302026
}\if{html}{\out{</div>}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Get the long-short ratio for any \code{\link[=available_tickers]{available_tickers()}} from the \code{\link[=available_exchanges]{available_exchanges()}}
}
\details{
\subsection{On time-zones and dates}{

Values passed to \code{from} or \code{to} must be coercible by \code{\link[=as.Date]{as.Date()}}, or \code{\link[=as.POSIXct]{as.POSIXct()}}, with a format of either \code{"\%Y-\%m-\%d"} or \code{"\%Y-\%m-\%d \%H:\%M:\%S"}. By default
all dates are passed and returned with \code{\link[=Sys.timezone]{Sys.timezone()}}.
}

\subsection{On returns}{

If only \code{from} is provided 200 pips are returned up to \code{Sys.time()}. If only \code{to} is provided 200 pips up to the specified date
is returned.
}
}
\note{
\subsection{Available exchanges}{

See \code{\link[=available_exchanges]{available_exchanges()}} with  for available exchanges.
}

\subsection{Limited return values}{

Binance only supports data for the last 30 days. Use other exchanges if you need beyond that.
}
}
\examples{
\dontrun{
  # script start;

  # 1) long-short ratio
  # on BTCUSDT pair
  ls_ratio <- cryptoQuotes::get_lsratio(
    ticker   = 'BTCUSDT',
    interval = '15m',
    from     = Sys.Date() - 1,
    to       = Sys.Date()
  )

  # 2) BTCSDT in same period
  # as the long-short ratio;
  BTC <- cryptoQuotes::get_quote(
    ticker   = 'BTCUSDT',
    futures  = TRUE,
    interval = '15m',
    from     = Sys.Date() - 1,
    to       = Sys.Date()
  )

  # 3) plot BTCUSDT-pair
  # with long-short ratio
  cryptoQuotes::chart(
    ticker = BTC,
    main   = cryptoQuotes::kline(),
    sub    = list(
      cryptoQuotes::lsr(ratio = ls_ratio),
      cryptoQuotes::volume()
    ),
    indicator = list(
      cryptoQuotes::bollinger_bands()
    )
  )
}



# end of scrtipt;
}
\seealso{
Other deprecated: 
\code{\link{availableExchanges}()},
\code{\link{availableIntervals}()},
\code{\link{availableTickers}()},
\code{\link{getFGIndex}()},
\code{\link{getQuote}()}
}
\author{
Jonas Cuzulan Hirani
}
\concept{deprecated}
