% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuote.R
\name{getQuote}
\alias{getQuote}
\title{Get a quote on a cryptopair from one
of the supported exchanges}
\usage{
getQuote(
 ticker,
 source   = 'binance',
 futures  = TRUE,
 interval = '1d',
 from     = NULL,
 to       = NULL
)
}
\arguments{
\item{ticker}{A character vector of length 1. Uppercase. See \code{\link[=availableTickers]{availableTickers()}} for available tickers.}

\item{source}{A character vector of length 1. See \code{\link[=availableExchanges]{availableExchanges()}} for available exchanges.}

\item{futures}{A logical value. Returns futures market if \link{TRUE}, spot market otherwise.}

\item{interval}{A character vector of length 1. See \code{\link[=availableIntervals]{availableIntervals()}} for available intervals.}

\item{from}{An optional character vector of length 1. Given in \%Y-\%m-\%d format.}

\item{to}{An optional character vector of length 1. Given in \%Y-\%m-\%d format.}
}
\value{
an xts object with Open, High, Low, Close and Volume. If \code{futures = TRUE} the prices are last prices.
}
\description{
Open, High, Low, Close, and Volume (OHLCV) quotes are essential pieces of
information used to analyze the price and trading activity of a financial asset over a specific time frame.
}
\details{
If only \code{from} is provided 100 pips are returned up to \code{Sys.Date()}.

If only \code{to} is provided 100 pips up to the specified date
is returned.

If \code{from} and \code{to} are both \link{NULL} 100 pips returned up to \code{Sys.Date()}
}
\examples{
# 1) load perpetual
# futures from Binance
# with 15m intervals
perpAtom <- try(
  cryptoQuotes::getQuote(
    ticker  = 'ATOMUSDT',
    source = 'binance',
    interval = '15m',
    futures = TRUE
  )
)

# 2) chart the futures
# using the chart
# function
if (!inherits(perpAtom, 'try-error')){
  cryptoQuotes::chart(
    chart = cryptoQuotes::kline(perpAtom) \%>\%
      cryptoQuotes::addVolume()   \%>\%
        cryptoQuotes::addBBands(cols = c('Close'))
  )
}

# NOTE: Without the try
# the examples fails on
# Github Actions


# script end;
}
