% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuote.R
\name{getQuote}
\alias{getQuote}
\title{Get the Open, High, Low, Close and Volume data on a cryptocurrency pair}
\usage{
## get OHLC-V
getQuote(
 ticker,
 source   = 'binance',
 futures  = TRUE,
 interval = '1d',
 from     = NULL,
 to       = NULL
)
}
\arguments{
\item{ticker}{An \link{character}-vector of \link{length} 1. See \code{\link[=available_tickers]{available_tickers()}} for available tickers.}

\item{source}{A \link{character}-vector of \link{length} 1. \code{binance} by default. See \code{\link[=available_exchanges]{available_exchanges()}} for available exchanges.}

\item{futures}{A \link{logical}-vector of \link{length} 1. \link{TRUE} by default. Returns futures market if \link{TRUE}, spot market otherwise.}

\item{interval}{A \link{character}-vector of \link{length} 1. \verb{1d} by default. See \code{\link[=available_intervals]{available_intervals()}} for available intervals.}

\item{from}{An optional \link{character}, \link{date} or \link{POSIXct} vector of \link{length} 1. \link{NULL} by default.}

\item{to}{An optional \link{character}, \link{date} or \link{POSIXct} vector of \link{length} 1. \link{NULL} by default.}
}
\value{
An \link{xts}-object containing,

\describe{
\item{}{<\link{POSIXct}> \strong{\code{index}:} The time-index}
\item{}{<\link{numeric}> \strong{\code{open}:} Opening price}
\item{}{<\link{numeric}> \strong{\code{high}:} Highest price}
\item{}{<\link{numeric}> \strong{\code{low}:} Lowest price}
\item{}{<\link{numeric}> \strong{\code{close}:} Closing price}
\item{}{<\link{numeric}> \strong{\code{volume}:} Trading volume}
}

\strong{Sample output}

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>                        open    high     low   close    volume
#> 2024-05-02 02:00:00 58298.5 59581.0 56887.3 59030.4 131444.90
#> 2024-05-03 02:00:00 59030.4 63343.7 58780.0 62843.7 150514.93
#> 2024-05-04 02:00:00 62843.7 64499.9 62512.8 63860.0  83894.05
#> 2024-05-05 02:00:00 63860.0 64630.0 62866.1 63997.8  70925.50
#> 2024-05-06 02:00:00 63997.8 65540.0 62669.0 63149.3 141354.74
#> 2024-05-07 02:00:00 63149.3 64000.0 62803.5 63774.3  27122.91
}\if{html}{\out{</div>}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Get a quote on a cryptocurrency pair from the \code{\link[=available_exchanges]{available_exchanges()}} in various \code{\link[=available_intervals]{available_intervals()}} for any actively traded \code{\link[=available_tickers]{available_tickers()}}.
}
\details{
\subsection{On time-zones and dates}{

Values passed to \code{from} or \code{to} must be coercible by \code{\link[=as.Date]{as.Date()}}, or \code{\link[=as.POSIXct]{as.POSIXct()}}, with a format of either \code{"\%Y-\%m-\%d"} or \code{"\%Y-\%m-\%d \%H:\%M:\%S"}. By default
all dates are passed and returned with \code{\link[=Sys.timezone]{Sys.timezone()}}.
}

\subsection{On returns}{

If only \code{from} is provided 200 pips are returned up to \code{Sys.time()}. If only \code{to} is provided 200 pips up to the specified date
is returned.
}
}
\examples{
\dontrun{
  # script start;

  # get quote on
  # BTCUSDT pair from
  # Binance in 30m
  # intervals from the
  # last 24 hours
  tail(
    BTC <- cryptoQuotes::get_quote(
      ticker   = 'BTCUSDT',
      source   = 'binance',
      interval = '30m',
      futures  = FALSE,
      from     = Sys.Date() - 1
    )
  )

  # script end;
}
}
\seealso{
Other deprecated: 
\code{\link{availableExchanges}()},
\code{\link{availableIntervals}()},
\code{\link{availableTickers}()},
\code{\link{getFGIndex}()},
\code{\link{getLSRatio}()}
}
\author{
Serkan Korkmaz
}
\concept{deprecated}
