% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuote.R
\name{getQuote}
\alias{getQuote}
\title{getQuote}
\usage{
getQuote(
  ticker,
  source = "binance",
  futures = TRUE,
  interval,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{ticker}{A character vector of length 1.}

\item{source}{A character vector of length 1.}

\item{futures}{A logical value. TRUE by default. If FALSE, the function will return spot prices.}

\item{interval}{A character vector of length 1.}

\item{from}{A character vector of length 1.}

\item{to}{A character vector of length 1.}
}
\description{
This function returns
a crypto quote
}
