% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuote.R
\name{getQuote}
\alias{getQuote}
\title{Get the Open, High, Low, Close and Volume data on a cryptocurrency pair}
\usage{
## get OHLC-V
getQuote(
 ticker,
 source   = 'binance',
 futures  = TRUE,
 interval = '1d',
 from     = NULL,
 to       = NULL
)
}
\arguments{
\item{ticker}{A \link{character}-vector of \link{length} 1. See \code{\link[=available_tickers]{available_tickers()}} for available tickers.}

\item{source}{A \link{character}-vector of \link{length} 1. \code{binance} by default. See \code{\link[=available_exchanges]{available_exchanges()}} for available exchanges.}

\item{futures}{A \link{logical}-vector of \link{length} 1. \link{TRUE} by default. Returns futures market if \link{TRUE}, spot market otherwise.}

\item{interval}{A \link{character}-vector of \link{length} 1. \verb{1d} by default. See \code{\link[=available_intervals]{available_intervals()}} for available intervals.}

\item{from}{An optional \link{character}-, \link{date}- or \link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}

\item{to}{An optional \link{character}-, \link{date}- or \link{POSIXct}-vector of \link{length} 1. \link{NULL} by default.}
}
\value{
An \link{xts}-object containing,

\item{index}{<\link{POSIXct}> The time-index}
\item{open}{<\link{numeric}> Opening price}
\item{high}{<\link{numeric}> Highest price}
\item{low}{<\link{numeric}> Lowest price}
\item{close}{<\link{numeric}> Closing price}
\item{volume}{<\link{numeric}> Trading volume}

\strong{Sample output}

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>                        open    high     low   close     volume
#> 2024-05-12 02:00:00 60810.2 61854.2 60566.6 61449.7  41972.384
#> 2024-05-13 02:00:00 61449.7 63437.3 60707.8 62909.4 105341.089
#> 2024-05-14 02:00:00 62909.4 63095.9 61000.1 61546.0 104281.757
#> 2024-05-15 02:00:00 61546.0 66430.0 61310.0 66185.4 143725.174
#> 2024-05-16 02:00:00 66185.4 66770.0 64580.4 65226.0  96849.243
#> 2024-05-17 02:00:00 65226.0 65868.0 65070.1 65449.9   9881.411
}\if{html}{\out{</div>}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Get a quote on a cryptocurrency pair from the \code{\link[=available_exchanges]{available_exchanges()}} in various \code{\link[=available_intervals]{available_intervals()}} for any actively traded \code{\link[=available_tickers]{available_tickers()}}.
}
\details{
\subsection{On time-zones and dates}{

Values passed to \code{from} or \code{to} must be coercible by \code{\link[=as.Date]{as.Date()}}, or \code{\link[=as.POSIXct]{as.POSIXct()}}, with a format of either \code{"\%Y-\%m-\%d"} or \code{"\%Y-\%m-\%d \%H:\%M:\%S"}. By default
all dates are passed and returned with \code{\link[=Sys.timezone]{Sys.timezone()}}.
}

\subsection{On returns}{

If only \code{from} is provided 200 pips are returned up to \code{Sys.time()}. If only \code{to} is provided 200 pips up to the specified date
is returned.
}
}
\examples{
\dontrun{
  # script start;

  # get quote on
  # BTCUSDT pair from
  # Binance in 30m
  # intervals from the
  # last 24 hours
  tail(
    BTC <- cryptoQuotes::get_quote(
      ticker   = 'BTCUSDT',
      source   = 'binance',
      interval = '30m',
      futures  = FALSE,
      from     = Sys.Date() - 1
    )
  )

  # script end;
}
}
\seealso{
Other deprecated: 
\code{\link{availableExchanges}()},
\code{\link{availableIntervals}()},
\code{\link{availableTickers}()},
\code{\link{getFGIndex}()},
\code{\link{getLSRatio}()}
}
\author{
Serkan Korkmaz
}
\concept{deprecated}
