% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_quote.R
\name{kline}
\alias{kline}
\title{Candlestick chart}
\usage{
kline(...)
}
\arguments{
\item{...}{For internal use. Please ignore.}
}
\value{
A \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A high-level \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-wrapper function for charting Open, High, Low and Close prices.
}
\examples{
# script start;

## charting the klines
## with indicators as
## subcharts
chart(
  ticker     = BTC,
  main       = kline(),
  sub        = list(
    volume(),
    macd()
  ),
  indicator = list(
    bollinger_bands(),
    sma(),
    alma()
  ),
  options = list(
    dark = TRUE,
    deficiency = FALSE
  )
)

# script end;
}
\seealso{
Other price charts: 
\code{\link{chart}()},
\code{\link{ohlc}()}
}
\author{
Serkan Korkmaz
}
\concept{price charts}
