% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_quote.R
\name{kline}
\alias{kline}
\title{Candlestick Chart}
\usage{
kline(...)
}
\arguments{
\item{...}{For internal use. Please ignore.}
}
\value{
An \link{invisible} \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A high-level \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-function for charting
Open, High, Low and Close prices.
}
\examples{
# script start;

# Charting BTC using
# candlesticks as main
# chart
cryptoQuotes::chart(
  ticker = BTC,
  main   = cryptoQuotes::kline(),
  sub    = list(
    cryptoQuotes::volume()
  )
)

# script end;
}
\seealso{
Other price charts: 
\code{\link{chart}()},
\code{\link{ohlc}()},
\code{\link{pline}()}
}
\author{
Serkan Korkmaz
}
\concept{price charts}
