% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_quote.R
\name{ohlc}
\alias{ohlc}
\title{Chart Open, High, Low and Close prices using OHLC-bars}
\usage{
ohlc(...)
}
\arguments{
\item{...}{For internal use. Please ignore.}
}
\value{
A \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A high-level \code{\link[plotly:plot_ly]{plotly::plot_ly()}}-function for charting Open, High, Low and Close prices.
}
\examples{
# script start;

# 1) charting weekly
# BTC using candlesticks
# and indicators
cryptoQuotes::chart(
  ticker     = BTC,
  main       = cryptoQuotes::kline(),
  sub        = list(
    cryptoQuotes::volume(),
    cryptoQuotes::macd()
  ),
  indicator = list(
    cryptoQuotes::bollinger_bands(),
    cryptoQuotes::sma(),
    cryptoQuotes::alma()
  ),
  options = list(
    dark       = TRUE,
    deficiency = FALSE
  )
)

# script end;
}
\seealso{
Other price charts: 
\code{\link{chart}()},
\code{\link{kline}()},
\code{\link{line}()}
}
\author{
Serkan Korkmaz
}
\concept{price charts}
